requirements.txt
django
djangorestframework
pytest
pytest-django